rename InfraCombinationView to InfraCombinationHistory
